Updated cfnresponse and requests modules
AWS is deprecating the use of botocore.requests which is still utilized
within the originally released zips (cfnresponse 1.0.2).  Upgrading to
cfnresponse 1.1.2 resolves the pending deprecation issue as
botocore.requests was replaced with urllib3.  It was also neccessary
to upgrade requests (2.23.0 -> 2.26.0) as the existing version of this
module had a legacy urllib3 dependency.