Add Dynamic Safe Identification for On-Boarding
This commit introduces greater control around which safe will be
targeted for on-boarding.  These new options include the use of an
EC2 Tag or an inferred account-level safe (particularly useful for
large multi-account deployments where accounts serve as a permission
boundary).

The safe selection workflow runs as follows:

1.)	Check the EC2 instance candidate for a Tag called “CyberArk-Safe”.
If this tag is found, lookup the defined value against PVWA REST API
to determine if the safe exists.  If the safe exists, use it for
on-boarding.  If the safe doesn’t exist (or the tag doesn’t exist),
cascade to step 2.

2.)	Arrange an inferable safe name of
“<Instance_Account_Id>_<AOB_*_Password>”.  For example, if the default
Unix safe is “AOB_Unix” and the instance’s AccountId is “123456789055”
then we infer a safe name of “123456789055_AOB_Unix”.  This safe name
is searched for via PVWA REST API.  If the safe exists, use it for
on-boarding.  If the safe doesn’t exist, cascade to step 3.

3.)	The default safe for Windows/Unix is used for on-boarding (current)

This approach allows for more refined safe control with each newly
introduced method remaining “optional”.  As such, these changes will
not disrupt existing implementations that are looking to implement
these optional safe controls for future on-boarding activity.