Fixed cfnresponse and Exception Handling
Refactored cfnresponse.Send() response_data parameter to
pass message in a dict; this will result in a more appropriate
message display in CFN if resource creation fails.

Added exception handling to delete_sessions_table.  This function
previously threw an unhandled exception if the table didn't exist
or would have otherwise encountered an error during the attempt.
This originally resulted in the custom resource getting stuck in
"Deleting" status for 60 minutes before finally failing.  If the table
does not exist, the function will now exit normally and delete tasks
will continue.  If there is an unexpected error, it will now result
in a proper cfnresponse.FAILED callback.