Improved Parameter Deletion Logic
If a parameter doesn't exist, it will no
longer cause an immediate cfnresponse.
This also removes the dependency on
get_aob_mode, which was removed.

Refactored deletion function to make it
more extensible.

Improved Parameter Deletion Logic

If a parameter doesn't exist, it will no
longer cause an immediate cfnresponse.
This also removes the dependency on
get_aob_mode, which was removed.

Refactored deletion function to make it
more extensible.