Updating CyberArk APIs to Gen 2 (Where Possible)
The following API calls have been updated from GEN 1 to GEN 2:

- Logon (raising the minimum PAS version requirement to v11.3)
- Logoff
- Add Account
- Get All Safes

The "Delete Account" API call is now the only call that remains at GEN 1, and this is due to the fact that the GEN 2 endpoint does not yet support SSH Key deletion.

The GEN 2 endpoint for Logon also includes an enhancing effect, as this endpoint does not require a consumer-managed connection number (1-100).  Instead, a "concurrentSession" flag is simply set to true in the request body, offloading concurrency management to the PVWA (with up to 300 concurrent sessions now also being supported).  As a result, this change also removes dependency on the DynamoDB Sessions table and accompanied session (connection number) acquisition logic.