Enhancing OS Distribution Username Identification
There are some marketplace AMIs that do not contain an image description, which was resulting in an on-boarding with an incorrect (for the target distribution) username of "ec2-user".  The "CIS Ubuntu Linux 20.04 LTS Benchmark - Level 1" AMI is an example of this.

Modified the image description setting logic such that if a description is not found on the image, use the image name instead (as this will in many cases contain an appropriate keyword for os distribution identification).